refactor: move autocommands and lsp attach logic to respective plugins
This commit refactors the location of certain autocommands and LSP
attach logic for better organization and maintainability. The
autocommands related to mode changes are moved to the auto-save plugin,
and the LSP attach logic is moved to the mason-lspconfig plugin. This
change improves the separation of concerns and makes the codebase
easier to understand and modify.